+++ pkg/shared/helldivers/lib/godivers.go
@@ -16,20 +16,20 @@ type Planet struct {
 	WaypointNames      string  `json:"waypointNames"`
 	Health             int     `json:"health"`
 	MaxHealth          int     `json:"maxHealth"`
-	Players            int     `json:"players"`
+	Players            int64   `json:"players"`
 	RegenPerSecond     float64 `json:"regenPerSecond"`
-	MissionsWon        int     `json:"missionsWon"`
-	MissionsLost       int     `json:"missionsLost"`
+	MissionsWon        int64   `json:"missionsWon"`
+	MissionsLost       int64   `json:"missionsLost"`
 	MissionTime        int64   `json:"missionTime"`
-	BugKills           int     `json:"bugKills"`
-	AutomatonKills     int     `json:"automatonKills"`
-	IlluminateKills    int     `json:"illuminateKills"`
+	BugKills           int64   `json:"bugKills"`
+	AutomatonKills     int64   `json:"automatonKills"`
+	IlluminateKills    int64   `json:"illuminateKills"`
 	BulletsFired       int64   `json:"bulletsFired"`
 	BulletsHit         int64   `json:"bulletsHit"`
 	TimePlayed         int64   `json:"timePlayed"`
-	Deaths             int     `json:"deaths"`
-	Revives            int     `json:"revives"`
-	Friendlies         int     `json:"friendlies"`
+	Deaths             int64   `json:"deaths"`
+	Revives            int64   `json:"revives"`
+	Friendlies         int64   `json:"friendlies"`
 	MissionSuccessRate int     `json:"missionSuccessRate"`
 	Accuracy           int     `json:"accurracy"`
 	LibPercent         float64 `json:"libPercent"`
